requirements.txt
streamlit 
rembg 
pillow
flask
